+++ ads-elk-etc/elastickibana_user-data.sh
@@ -1,40 +1,22 @@
 #!/bin/bash
 
 # Install docker and git
-yum install docker git -y
-
-# Clone the repository that contains the relevant docker files
-git clone https://github.com/adsabs/aws_elk.git
+yum install docker -y
 
 # Start docker
 service docker start
 
-# Build Elasticsearch and Kibana containers
-pushd aws_elk/
-docker build --tag adsabs/elasticsearch Dockerfiles/elasticsearch/
-docker build --tag adsabs/kibana Dockerfiles/kibana/
-
-# Copy the config file
-mkdir -p /etc/elasticsearch/
-pushd /etc/elasticsearch/
-aws s3 cp s3://adsabs-elk-etc/elasticsearch.yml elasticsearch.yml
-aws s3 cp s3://adsabs-elk-etc/logging.yml logging.yml
-popd
-
-mkdir -p /etc/kibana/config/
-pushd /etc/kibana/config/
-aws s3 cp s3://adsabs-elk-etc/kibana.yml kibana.yml
-popd
-
-
 # Run the containers
 ## First start elasticsearch
-docker run -d --name elasticsearch -p 9200:9200 -p 9300:9300 -v /etc/elasticsearch/:/etc/elasticsearch/ adsabs/elasticsearch
+docker run -d --name elasticsearch -p 9200:9200 -p 9300:9300 adsabs/elasticsearch:d7cd21774e7a5c62c82af77871b2f3fec31f9506
 
 ## Second start kibana
 ## Kibana should wait for elasticsearch to be ready, otherwise it will crash
 ## But I do not want it to stay in an infinite loop, so after 60 seconds, there is
 ## clearly something wrong with elasticsearch that warrants users intervention
+
+## Given phusion/image is used for kibana, it should be fine that you do not use
+## the following, as it will keep trying to restart, but for now I will leave it.
 
 ELASTICSEARCH_HOST="http://localhost:9200"
 time_out=60
@@ -58,6 +40,5 @@ do
   sleep 1
 done
 
-docker run -d --name kibana --link elasticsearch:elasticsearch -p 5601:5601 -v /etc/kibana/config/kibana.yml:/etc/kibana/config/kibana.yml adsabs/kibana
+docker run -d --name kibana --link elasticsearch:elasticsearch -p 5601:5601 adsabs/kibana:cdf22333579930dd777cc63ce779658fccd45b7a
 
-popd